+++ device/device.go
@@ -1,0 +1,8 @@
+package device
+
+type model struct {
+	name        string
+	description string
+	mac_address string
+	ip_address  string
+}
